fix(trusty-mpm): complete PM-guard shell separator set (bare &, newline) + bound substitution recursion (review 2)
Second adversarial review found two remaining issues:
- CRITICAL: split_shell_segments missed the bare `&` background separator, so
  `true & sed -i x f` bypassed the deny check. Complete the separator set to the
  full POSIX set (&&, ||, ;, |, &, newline); the bare `&` splits except when it's
  a redirection fd-dup (`2>&1`, `>&2`, `&>file`) or trailing background (`foo &`).
- MEDIUM: unbounded $(...)/backtick recursion could stack-overflow the hook on
  adversarial nesting; thread an explicit depth, cap at 32, deny conservatively
  past the cap.

Extracted the bash classifier into pm_guard_bash.rs (398 SLOC) to keep both files
under the 500-SLOC cap (pm_guard.rs now 145). +18 regression tests (bare-&
deny, trailing-bg allow, fd-dup not split, deep-nesting bounded). fmt/clippy/
line-cap clean; cargo test -p trusty-mpm green.
